ins-001/api/requirements.txt
# INS-001 Semantic Associations - Backend Dependencies
# Pin ALL versions to prevent agent hallucination

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
# Upgraded to newer version compatible with httpx 0.26+
supabase>=2.3.4
# Let supabase manage its own dependencies

# OpenAI
# Upgraded to newer version compatible with httpx 0.26+
openai>=1.12.0

# Anthropic
anthropic==0.42.0

# HTTP client (for backup job)
# Newer supabase versions support httpx 0.26+
httpx>=0.26,<0.29

# Note: asyncio is built into Python 3.4+, no need to install

# Math/vectors
numpy==1.26.3

# Retry logic
tenacity==8.2.3

# Environment
python-dotenv==1.0.1

# Type hints
pydantic==2.6.0

# Analytics
posthog==3.3.1
sentry-sdk[fastapi]==1.40.0

# Vocabulary loading
wordfreq==3.0.2

# Dev
# Note: pytest-asyncio 0.23.4 requires pytest<8
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
